requirements.txt
pygame>=2.0.0
pyrebase4>=4.5.0
